rename to requestauthplugin
